+++ packages/components/src/components/SquiggleViewer/SquiggleValueHeader.tsx
@@ -1,0 +1,18 @@
+import { FC } from "react";
+
+import { SqValueWithContext } from "../../lib/utility.js";
+import { widgetRegistry } from "../widgets/registry.js";
+
+export const SquiggleValueHeader: FC<{
+  value: SqValueWithContext;
+}> = ({ value }) => {
+  const widget = widgetRegistry.widgets.get(value.tag);
+
+  const heading = widget?.heading?.(value) || value.publicName();
+
+  return (
+    <div className="text-stone-400 group-hover:text-stone-600 text-sm transition">
+      {heading}
+    </div>
+  );
+};
