+++ packages/components/src/components/widgets/BoolWidget.tsx
@@ -1,0 +1,17 @@
+import { clsx } from "clsx";
+import { widgetRegistry } from "./registry.js";
+import { leftWidgetMargin } from "./utils.js";
+
+widgetRegistry.register("Bool", {
+  renderPreview: (value) => value.value.toString(),
+  render: (value) => (
+    <div
+      className={clsx(
+        "text-indigo-800 text-sm font-mono font-semibold",
+        leftWidgetMargin
+      )}
+    >
+      {value.value.toString()}
+    </div>
+  ),
+});
