+++ packages/components/src/components/widgets/DateWidget.tsx
@@ -1,0 +1,5 @@
+import { widgetRegistry } from "./registry.js";
+
+widgetRegistry.register("Date", {
+  render: (value) => value.value.toDateString(),
+});
