+++ packages/components/src/components/widgets/PlotWidget.tsx
@@ -1,0 +1,61 @@
+import { DistributionsChart } from "../DistributionsChart/index.js";
+import { DistFunctionChart } from "../FunctionChart/DistFunctionChart.js";
+import { NumericFunctionChart } from "../FunctionChart/NumericFunctionChart.js";
+import { RelativeValuesGridChart } from "../RelativeValuesGridChart/index.js";
+import { ScatterChart } from "../ScatterChart/index.js";
+import { CHART_TO_DIST_HEIGHT_ADJUSTMENT } from "./DistWidget.js";
+import { widgetRegistry } from "./registry.js";
+
+widgetRegistry.register("Plot", {
+  render: (value, settings) => {
+    const plot = value.value;
+    const environment = value.context.project.getEnvironment();
+
+    switch (plot.tag) {
+      case "distributions":
+        return (
+          <DistributionsChart
+            plot={plot}
+            environment={environment}
+            height={settings.chartHeight * CHART_TO_DIST_HEIGHT_ADJUSTMENT}
+          />
+        );
+      case "numericFn": {
+        return (
+          <NumericFunctionChart
+            plot={plot}
+            environment={environment}
+            height={settings.chartHeight}
+          />
+        );
+      }
+      case "distFn": {
+        return (
+          <DistFunctionChart
+            plot={plot}
+            environment={{
+              sampleCount: environment.sampleCount / 10,
+              xyPointLength: environment.xyPointLength / 10,
+            }}
+            height={settings.chartHeight}
+          />
+        );
+      }
+      case "scatter":
+        return (
+          <ScatterChart
+            plot={plot}
+            environment={environment}
+            height={settings.chartHeight}
+          />
+        );
+      case "relativeValues":
+        return (
+          <RelativeValuesGridChart plot={plot} environment={environment} />
+        );
+      default:
+        // can happen if squiggle-lang version is too fresh and we messed up the components -> squiggle-lang dependency
+        return `Unsupported plot ${plot satisfies never}`;
+    }
+  },
+});
