+++ packages/components/src/components/widgets/SqTypeWithCount.tsx
@@ -1,0 +1,11 @@
+import { FC } from "react";
+
+export const SqTypeWithCount: FC<{
+  type: string;
+  count: number;
+}> = ({ type, count }) => (
+  <div>
+    {type}
+    <span className="ml-0.5">{count}</span>
+  </div>
+);
