+++ packages/components/src/components/widgets/TimeDurationWidget.tsx
@@ -1,0 +1,6 @@
+import { NumberShower } from "../NumberShower.js";
+import { widgetRegistry } from "./registry.js";
+
+widgetRegistry.register("TimeDuration", {
+  render: (value) => <NumberShower precision={3} number={value.value} />,
+});
